+++ src/render/canvas_helper.rs
@@ -50,4 +50,14 @@ pub trait CanvasHelper: Widget {
             .unwrap();
         }
     }
+
+    /// Returns a `Rect` destination object
+    fn get_rect_dest(&mut self) -> Rect {
+        Rect::new(
+            self.get_config().to_x(0),
+            self.get_config().to_y(0),
+            self.get_config().get_size(CONFIG_SIZE)[SIZE_WIDTH],
+            self.get_config().get_size(CONFIG_SIZE)[SIZE_HEIGHT],
+        )
+    }
 }