+++ src/widgets/checkbox_widget.rs
@@ -65,28 +65,28 @@ impl CheckboxWidget {
             sdl2::ttf::FontStyle::NORMAL,
             font_size,
             TextJustify::Left,
-            text.clone(),
+            text,
             make_points(
-                points[POINT_X].clone() + size[SIZE_HEIGHT].clone() as i32 + 6,
-                points[POINT_Y].clone() + 2,
+                points[POINT_X] + size[SIZE_HEIGHT] as i32 + 6,
+                points[POINT_Y] + 2,
             ),
             make_size(
-                size[SIZE_WIDTH].clone() - size[SIZE_HEIGHT].clone() - 10,
-                size[SIZE_HEIGHT].clone() - 4,
+                size[SIZE_WIDTH] - size[SIZE_HEIGHT] - 10,
+                size[SIZE_HEIGHT] - 4,
             ),
         );
 
         let mut config = WidgetConfig::new(points.clone(), size.clone());
         let mut unchecked_widget = ImageWidget::new(
             String::from("assets/checkbox_unselected.png"),
-            make_points(points[POINT_X].clone() + 2, points[POINT_Y].clone() + 2),
-            make_size(size[SIZE_HEIGHT].clone() - 4, size[SIZE_HEIGHT].clone() - 4),
+            make_points(points[POINT_X] + 2, points[POINT_Y] + 2),
+            make_size(size[SIZE_HEIGHT] - 4, size[SIZE_HEIGHT] - 4),
             true,
         );
         let mut checked_widget = ImageWidget::new(
             String::from("assets/checkbox_selected.png"),
-            make_points(points[POINT_X].clone() + 2, points[POINT_Y].clone() + 2),
-            make_size(size[SIZE_HEIGHT].clone() - 4, size[SIZE_HEIGHT].clone() - 4),
+            make_points(points[POINT_X] + 2, points[POINT_Y] + 2),
+            make_size(size[SIZE_HEIGHT] - 4, size[SIZE_HEIGHT] - 4),
             true,
         );